@@ -9,18 +9,10 @@ dependencies:
   flutter_svg:
     path: ../
 
-# For information on the generic Dart part of this file, see the
-# following page: https://www.dartlang.org/tools/pub/pubspec
-
 # The following section is specific to Flutter.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
 
-  # To add assets to your application, add an assets section, like this:
   assets:
     - assets/
     - assets/w3samples/
@@ -30,8 +22,8 @@ flutter:
     - assets/android_vd/
 
 environment:
-  sdk: ">=1.19.0 <3.0.0"
-  flutter: ">=0.7.3 <2.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
+  flutter: ">=1.24.0-6.0.pre"
 
 # dependency_overrides:
 #  sky_engine:
