@@ -1,26 +1,26 @@
 name: flutter_svg
 description: An SVG rendering and widget library for Flutter, which allows painting and displaying Scalable Vector Graphics 1.1 files.
 homepage: https://github.com/dnfield/flutter_svg
-version: 0.19.1+2
+version: 0.20.0-nullsafety.4+1
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.3.0-nullsafety.3
+  meta: ^1.3.0
   path_drawing: ^0.4.1+1
-  vector_math: ^2.1.0-nullsafety.3
-  xml: ^4.5.1
+  vector_math: ^2.1.0
+
+  xml: ^4.2.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.1
-  path: ^1.7.0
-  test: ^1.14.3
+  path: ^1.8.0
+  test: ^1.16.0
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
-  flutter: ">=1.22.0 <2.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
+  flutter: ">=1.24.0-6.0.pre"
 
 ## For developing features that require new functionality in engine:
 # dependency_overrides:
