@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// ignore_for_file: deprecated_member_use_from_same_package
-
 import 'dart:async';
 import 'dart:io';
 import 'dart:typed_data';
@@ -19,8 +17,15 @@ import 'picture_cache.dart';
 import 'picture_stream.dart';
 import 'utilities/http.dart';
 
+/// The signature of a function that can decode raw SVG data into a [Picture].
+///
+/// Used by [PictureProvider]. Several useful methods are provided in the [Svg]
+/// class.
 typedef PictureInfoDecoder<T> = Future<PictureInfo> Function(
-    T data, ColorFilter colorFilter, String key);
+  T data,
+  ColorFilter? colorFilter,
+  String key,
+);
 
 /// Creates an [PictureConfiguration] based on the given [BuildContext] (and
 /// optionally size).
@@ -36,21 +41,20 @@ typedef PictureInfoDecoder<T> = Future<PictureInfo> Function(
 ///
 ///  * [PictureProvider], which has an example showing how this might be used.
 PictureConfiguration createLocalPictureConfiguration(
-  BuildContext context, {
-  Rect viewBox,
-  ColorFilter colorFilterOverride,
-  Color color,
-  BlendMode colorBlendMode,
+  BuildContext? context, {
+  Rect? viewBox,
+  ColorFilter? colorFilterOverride,
+  Color? color,
+  BlendMode? colorBlendMode,
 }) {
-  ColorFilter filter = colorFilterOverride;
+  ColorFilter? filter = colorFilterOverride;
   if (filter == null && color != null) {
     filter = ColorFilter.mode(color, colorBlendMode ?? BlendMode.srcIn);
   }
   return PictureConfiguration(
     bundle: context != null ? DefaultAssetBundle.of(context) : rootBundle,
-    locale:
-        context != null ? Localizations.localeOf(context, nullOk: true) : null,
-    textDirection: context != null ? Directionality.of(context) : null,
+    locale: context != null ? Localizations.maybeLocaleOf(context) : null,
+    textDirection: context != null ? Directionality.maybeOf(context) : null,
     viewBox: viewBox,
     platform: defaultTargetPlatform,
     colorFilter: filter,
@@ -78,7 +82,7 @@ class PictureConfiguration {
     this.textDirection,
     this.viewBox,
     this.platform,
-    @deprecated this.colorFilter,
+    this.colorFilter,
   });
 
   /// Creates an object holding the configuration information for an [PictureProvider].
@@ -86,12 +90,12 @@ class PictureConfiguration {
   /// All the arguments are optional. Configuration information is merely
   /// advisory and best-effort.
   PictureConfiguration copyWith({
-    AssetBundle bundle,
-    Locale locale,
-    TextDirection textDirection,
-    Rect viewBox,
-    TargetPlatform platform,
-    ColorFilter colorFilter,
+    AssetBundle? bundle,
+    Locale? locale,
+    TextDirection? textDirection,
+    Rect? viewBox,
+    TargetPlatform? platform,
+    ColorFilter? colorFilter,
   }) {
     return PictureConfiguration(
       bundle: bundle ?? this.bundle,
@@ -105,27 +109,25 @@ class PictureConfiguration {
 
   /// The preferred [AssetBundle] to use if the [PictureProvider] needs one and
   /// does not have one already selected.
-  final AssetBundle bundle;
+  final AssetBundle? bundle;
 
   /// The language and region for which to select the picture.
-  final Locale locale;
+  final Locale? locale;
 
   /// The reading direction of the language for which to select the picture.
-  final TextDirection textDirection;
+  final TextDirection? textDirection;
 
   /// The size at which the picture will be rendered.
-  final Rect viewBox;
+  final Rect? viewBox;
 
   /// The [TargetPlatform] for which assets should be used. This allows pictures
   /// to be specified in a platform-neutral fashion yet use different assets on
   /// different platforms, to match local conventions e.g. for color matching or
   /// shadows.
-  final TargetPlatform platform;
+  final TargetPlatform? platform;
 
   /// The [ColorFilter], if any, that was applied to the drawing.
-  @Deprecated(
-      'Use a ColorFilter widget instead, or the color parameter on SvgPicture.')
-  final ColorFilter colorFilter;
+  final ColorFilter? colorFilter;
 
   /// a picture configuration that provides no additional information.
   ///
@@ -156,9 +158,6 @@ class PictureConfiguration {
     result.write('PictureConfiguration(');
     bool hasArguments = false;
     if (bundle != null) {
-      if (hasArguments) {
-        result.write(', ');
-      }
       result.write('bundle: $bundle');
       hasArguments = true;
     }
@@ -187,7 +186,7 @@ class PictureConfiguration {
       if (hasArguments) {
         result.write(', ');
       }
-      result.write('platform: ${describeEnum(platform)}');
+      result.write('platform: ${describeEnum(platform!)}');
       hasArguments = true;
     }
     if (colorFilter != null) {
@@ -303,13 +302,16 @@ PictureCache _cache = PictureCache();
 abstract class PictureProvider<T> {
   /// Abstract const constructor. This constructor enables subclasses to provide
   /// const constructors so that they can be used in const expressions.
-  const PictureProvider();
+  const PictureProvider(this.colorFilter);
 
   /// The number of items in the [PictureCache].
   static int get cacheCount => _cache.count;
 
   /// Clears the [PictureCache].
   static void clearCache() => _cache.clear();
+
+  /// The color filter to apply to the picture, if any.
+  final ColorFilter? colorFilter;
 
   /// Resolves this Picture provider using the given `configuration`, returning
   /// an [PictureStream].
@@ -319,22 +321,25 @@ abstract class PictureProvider<T> {
   /// Subclasses should implement [obtainKey] and [load], which are used by this
   /// method.
   PictureStream resolve(PictureConfiguration picture,
-      {PictureErrorListener onError}) {
+      {PictureErrorListener? onError}) {
+    // ignore: unnecessary_null_comparison
     assert(picture != null);
     final PictureStream stream = PictureStream();
-    T obtainedKey;
-    obtainKey(picture).then<void>((T key) {
-      obtainedKey = key;
-      stream.setCompleter(
-        _cache.putIfAbsent(
-          key,
-          () => load(key, onError: onError),
-        ),
-      );
-    }).catchError((dynamic exception, StackTrace stack) async {
+    T? obtainedKey;
+    obtainKey(picture).then<void>(
+      (T key) {
+        obtainedKey = key;
+        stream.setCompleter(
+          _cache.putIfAbsent(
+            key!,
+            () => load(key, onError: onError),
+          ),
+        );
+      },
+    ).catchError((Object exception, StackTrace stack) async {
       if (onError != null) {
         onError(exception, stack);
-        return null;
+        return;
       }
       FlutterError.reportError(FlutterErrorDetails(
           exception: exception,
@@ -348,7 +353,6 @@ abstract class PictureProvider<T> {
             yield DiagnosticsProperty<T>('Picture key', obtainedKey,
                 defaultValue: null);
           }));
-      return null;
     });
     return stream;
   }
@@ -367,7 +371,7 @@ abstract class PictureProvider<T> {
   /// Converts a key into an [PictureStreamCompleter], and begins fetching the
   /// picture.
   @protected
-  PictureStreamCompleter load(T key, {PictureErrorListener onError});
+  PictureStreamCompleter load(T key, {PictureErrorListener? onError});
 
   @override
   String toString() => '$runtimeType()';
@@ -382,11 +386,9 @@ class AssetBundlePictureKey {
   ///
   /// The arguments must not be null.
   const AssetBundlePictureKey(
-      {@required this.bundle,
-      @required this.name,
-      @deprecated this.colorFilter})
-      : assert(bundle != null),
-        assert(name != null);
+      {required this.bundle, required this.name, this.colorFilter})
+      : assert(bundle != null), // ignore: unnecessary_null_comparison
+        assert(name != null); // ignore: unnecessary_null_comparison
 
   /// The bundle from which the picture will be obtained.
   ///
@@ -399,9 +401,7 @@ class AssetBundlePictureKey {
   final String name;
 
   /// The [ColorFilter], if any, to be applied to the drawing.
-  @Deprecated(
-      'Use a ColorFilter widget instead, or the color parameter on SvgPicture.')
-  final ColorFilter colorFilter;
+  final ColorFilter? colorFilter;
 
   @override
   bool operator ==(dynamic other) {
@@ -430,7 +430,9 @@ abstract class AssetBundlePictureProvider
     extends PictureProvider<AssetBundlePictureKey> {
   /// Abstract const constructor. This constructor enables subclasses to provide
   /// const constructors so that they can be used in const expressions.
-  const AssetBundlePictureProvider(this.decoder) : assert(decoder != null);
+  const AssetBundlePictureProvider(this.decoder, ColorFilter? colorFilter)
+      : assert(decoder != null), // ignore: unnecessary_null_comparison
+        super(colorFilter);
 
   /// The decoder to use to turn a string into a [PictureInfo] object.
   final PictureInfoDecoder<String> decoder;
@@ -439,7 +441,7 @@ abstract class AssetBundlePictureProvider
   /// picture using [_loadAsync].
   @override
   PictureStreamCompleter load(AssetBundlePictureKey key,
-      {PictureErrorListener onError}) {
+      {PictureErrorListener? onError}) {
     return OneFramePictureStreamCompleter(_loadAsync(key, onError),
         informationCollector: () sync* {
       yield DiagnosticsProperty<PictureProvider>('Picture provider', this);
@@ -453,14 +455,17 @@ abstract class AssetBundlePictureProvider
   /// This function is used by [load].
   @protected
   Future<PictureInfo> _loadAsync(
-      AssetBundlePictureKey key, PictureErrorListener onError) async {
+      AssetBundlePictureKey key, PictureErrorListener? onError) async {
     final String data = await key.bundle.loadString(key.name);
-    if (data == null) {
-      throw 'Unable to read data';
-    }
     if (onError != null) {
-      return decoder(data, key.colorFilter, key.toString())
-        ..catchError(onError);
+      return decoder(
+        data,
+        key.colorFilter,
+        key.toString(),
+      ).catchError((Object error, StackTrace stack) {
+        onError(error, stack);
+        return Future<PictureInfo>.error(error, stack);
+      });
     }
     return decoder(data, key.colorFilter, key.toString());
   }
@@ -481,8 +486,9 @@ class NetworkPicture extends PictureProvider<NetworkPicture> {
   ///
   /// The arguments must not be null.
   const NetworkPicture(this.decoder, this.url,
-      {this.headers, @deprecated this.colorFilter})
-      : assert(url != null);
+      {this.headers, ColorFilter? colorFilter})
+      : assert(url != null), // ignore: unnecessary_null_comparison
+        super(colorFilter);
 
   /// The decoder to use to turn a [Uint8List] into a [PictureInfo] object.
   final PictureInfoDecoder<Uint8List> decoder;
@@ -491,12 +497,7 @@ class NetworkPicture extends PictureProvider<NetworkPicture> {
   final String url;
 
   /// The HTTP headers that will be used with [HttpClient.get] to fetch picture from network.
-  final Map<String, String> headers;
-
-  /// The [ColorFilter], if any, to apply to the drawing.
-  @Deprecated(
-      'Use a ColorFilter widget instead, or the color parameter on SvgPicture.')
-  final ColorFilter colorFilter;
+  final Map<String, String>? headers;
 
   @override
   Future<NetworkPicture> obtainKey(PictureConfiguration picture) {
@@ -505,7 +506,7 @@ class NetworkPicture extends PictureProvider<NetworkPicture> {
 
   @override
   PictureStreamCompleter load(NetworkPicture key,
-      {PictureErrorListener onError}) {
+      {PictureErrorListener? onError}) {
     return OneFramePictureStreamCompleter(_loadAsync(key, onError: onError),
         informationCollector: () sync* {
       yield DiagnosticsProperty<PictureProvider>('Picture provider', this);
@@ -514,11 +515,18 @@ class NetworkPicture extends PictureProvider<NetworkPicture> {
   }
 
   Future<PictureInfo> _loadAsync(NetworkPicture key,
-      {PictureErrorListener onError}) async {
+      {PictureErrorListener? onError}) async {
     assert(key == this);
-    final Uint8List bytes = await httpGet(url);
+    final Uint8List bytes = await httpGet(url, headers: headers);
     if (onError != null) {
-      return decoder(bytes, colorFilter, key.toString())..catchError(onError);
+      return decoder(
+        bytes,
+        colorFilter,
+        key.toString(),
+      ).catchError((Object error, StackTrace stack) {
+        onError(error, stack);
+        return Future<PictureInfo>.error(error, stack);
+      });
     }
     return decoder(bytes, colorFilter, key.toString());
   }
@@ -551,9 +559,10 @@ class FilePicture extends PictureProvider<FilePicture> {
   /// Creates an object that decodes a [File] as a picture.
   ///
   /// The arguments must not be null.
-  const FilePicture(this.decoder, this.file, {@deprecated this.colorFilter})
-      : assert(decoder != null),
-        assert(file != null);
+  const FilePicture(this.decoder, this.file, {ColorFilter? colorFilter})
+      : assert(decoder != null), // ignore: unnecessary_null_comparison
+        assert(file != null), // ignore: unnecessary_null_comparison
+        super(colorFilter);
 
   /// The file to decode into a picture.
   final File file;
@@ -561,34 +570,37 @@ class FilePicture extends PictureProvider<FilePicture> {
   /// The [PictureInfoDecoder] to use for loading this picture.
   final PictureInfoDecoder<Uint8List> decoder;
 
-  /// The [ColorFilter], if any, to use when drawing this picture.
-  @Deprecated(
-      'Use a ColorFilter widget instead, or the color parameter on SvgPicture.')
-  final ColorFilter colorFilter;
-
   @override
   Future<FilePicture> obtainKey(PictureConfiguration picture) {
     return SynchronousFuture<FilePicture>(this);
   }
 
   @override
-  PictureStreamCompleter load(FilePicture key, {PictureErrorListener onError}) {
+  PictureStreamCompleter load(FilePicture key,
+      {PictureErrorListener? onError}) {
     return OneFramePictureStreamCompleter(_loadAsync(key, onError: onError),
         informationCollector: () sync* {
-      yield DiagnosticsProperty<String>('Path', file?.path);
+      yield DiagnosticsProperty<String>('Path', file.path);
     });
   }
 
-  Future<PictureInfo> _loadAsync(FilePicture key,
-      {PictureErrorListener onError}) async {
+  Future<PictureInfo?> _loadAsync(FilePicture key,
+      {PictureErrorListener? onError}) async {
     assert(key == this);
 
     final Uint8List data = await file.readAsBytes();
-    if (data == null || data.isEmpty) {
+    if (data.isEmpty) {
       return null;
     }
     if (onError != null) {
-      return decoder(data, colorFilter, key.toString())..catchError(onError);
+      return decoder(
+        data,
+        colorFilter,
+        key.toString(),
+      ).catchError((Object error, StackTrace stack) async {
+        onError(error, stack);
+        return Future<PictureInfo>.error(error, stack);
+      });
     }
     return decoder(data, colorFilter, key.toString());
   }
@@ -599,16 +611,16 @@ class FilePicture extends PictureProvider<FilePicture> {
       return false;
     }
     return other is FilePicture &&
-        file?.path == other.file?.path &&
+        file.path == other.file.path &&
         other.colorFilter == colorFilter;
   }
 
   @override
-  int get hashCode => hashValues(file?.path?.hashCode, colorFilter);
+  int get hashCode => hashValues(file.path.hashCode, colorFilter);
 
   @override
   String toString() =>
-      '$runtimeType("${file?.path}", colorFilter: $colorFilter)';
+      '$runtimeType("${file.path}", colorFilter: $colorFilter)';
 }
 
 /// Decodes the given [String] buffer as a picture, associating it with the
@@ -627,16 +639,12 @@ class MemoryPicture extends PictureProvider<MemoryPicture> {
   /// Creates an object that decodes a [Uint8List] buffer as a picture.
   ///
   /// The arguments must not be null.
-  const MemoryPicture(this.decoder, this.bytes, {@deprecated this.colorFilter})
-      : assert(bytes != null);
+  const MemoryPicture(this.decoder, this.bytes, {ColorFilter? colorFilter})
+      : assert(bytes != null), // ignore: unnecessary_null_comparison
+        super(colorFilter);
 
   /// The [PictureInfoDecoder] to use when drawing this picture.
   final PictureInfoDecoder<Uint8List> decoder;
-
-  /// The [ColorFilter], if any, to use when drawing this picture.
-  @Deprecated(
-      'Use a ColorFilter widget instead, or the color parameter on SvgPicture.')
-  final ColorFilter colorFilter;
 
   /// The bytes to decode into a picture.
   final Uint8List bytes;
@@ -648,15 +656,22 @@ class MemoryPicture extends PictureProvider<MemoryPicture> {
 
   @override
   PictureStreamCompleter load(MemoryPicture key,
-      {PictureErrorListener onError}) {
+      {PictureErrorListener? onError}) {
     return OneFramePictureStreamCompleter(_loadAsync(key, onError: onError));
   }
 
   Future<PictureInfo> _loadAsync(MemoryPicture key,
-      {PictureErrorListener onError}) async {
+      {PictureErrorListener? onError}) async {
     assert(key == this);
     if (onError != null) {
-      return decoder(bytes, colorFilter, key.toString())..catchError(onError);
+      return decoder(
+        bytes,
+        colorFilter,
+        key.toString(),
+      ).catchError((Object error, StackTrace stack) {
+        onError(error, stack);
+        return Future<PictureInfo>.error(error, stack);
+      });
     }
     return decoder(bytes, colorFilter, key.toString());
   }
@@ -694,16 +709,12 @@ class StringPicture extends PictureProvider<StringPicture> {
   /// Creates an object that decodes a [Uint8List] buffer as a picture.
   ///
   /// The arguments must not be null.
-  const StringPicture(this.decoder, this.string, {@deprecated this.colorFilter})
-      : assert(string != null);
+  const StringPicture(this.decoder, this.string, {ColorFilter? colorFilter})
+      : assert(string != null), // ignore: unnecessary_null_comparison
+        super(colorFilter);
 
   /// The [PictureInfoDecoder] to use for decoding this picture.
   final PictureInfoDecoder<String> decoder;
-
-  /// The [ColorFilter], if any, to use when drawing this picture.
-  @Deprecated(
-      'Use a ColorFilter widget instead, or the color parameter on SvgPicture.')
-  final ColorFilter colorFilter;
 
   /// The string to decode into a picture.
   final String string;
@@ -715,17 +726,24 @@ class StringPicture extends PictureProvider<StringPicture> {
 
   @override
   PictureStreamCompleter load(StringPicture key,
-      {PictureErrorListener onError}) {
+      {PictureErrorListener? onError}) {
     return OneFramePictureStreamCompleter(_loadAsync(key, onError: onError));
   }
 
   Future<PictureInfo> _loadAsync(
     StringPicture key, {
-    PictureErrorListener onError,
+    PictureErrorListener? onError,
   }) {
     assert(key == this);
     if (onError != null) {
-      return decoder(string, colorFilter, key.toString())..catchError(onError);
+      return decoder(
+        string,
+        colorFilter,
+        key.toString(),
+      ).catchError((Object error, StackTrace stack) {
+        onError(error, stack);
+        return Future<PictureInfo>.error(error, stack);
+      });
     }
     return decoder(string, colorFilter, key.toString());
   }
@@ -834,9 +852,9 @@ class ExactAssetPicture extends AssetBundlePictureProvider {
     this.assetName, {
     this.bundle,
     this.package,
-    @deprecated this.colorFilter,
-  })  : assert(assetName != null),
-        super(decoder);
+    ColorFilter? colorFilter,
+  })  : assert(assetName != null), // ignore: unnecessary_null_comparison
+        super(decoder, colorFilter);
 
   /// The name of the asset.
   final String assetName;
@@ -846,11 +864,6 @@ class ExactAssetPicture extends AssetBundlePictureProvider {
   String get keyName =>
       package == null ? assetName : 'packages/$package/$assetName';
 
-  /// The [ColorFilter], if any, to use when drawing this picture.
-  @Deprecated(
-      'Use a ColorFilter widget instead, or the color parameter on SvgPicture.')
-  final ColorFilter colorFilter;
-
   /// The bundle from which the picture will be obtained.
   ///
   /// If the provided [bundle] is null, the bundle provided in the
@@ -859,11 +872,11 @@ class ExactAssetPicture extends AssetBundlePictureProvider {
   ///
   /// The picture is obtained by calling [AssetBundle.load] on the given [bundle]
   /// using the key given by [keyName].
-  final AssetBundle bundle;
+  final AssetBundle? bundle;
 
   /// The name of the package from which the picture is included. See the
   /// documentation for the [ExactAssetPicture] class itself for details.
-  final String package;
+  final String? package;
 
   @override
   Future<AssetBundlePictureKey> obtainKey(PictureConfiguration picture) {
