@@ -8,7 +8,20 @@ class SvgParser {
   /// Parses SVG from a string to a [DrawableRoot].
   ///
   /// The [key] parameter is used for debugging purposes.
-  Future<DrawableRoot> parse(String str, {String key}) async {
-    return await SvgParserState(xml.parseEvents(str), key).parse();
+  ///
+  /// By default SVG parsing will only log warnings when detecting unsupported
+  /// elements in an SVG.
+  /// If [warningsAsErrors] is true the function will throw with an error
+  /// instead.
+  /// You might want to set this to true for test and to false at runtime.
+  /// Defaults to false.
+  Future<DrawableRoot> parse(
+    String str, {
+    String? key,
+    bool warningsAsErrors = false,
+  }) async {
+    final SvgParserState state =
+        SvgParserState(xml.parseEvents(str), key, warningsAsErrors);
+    return await state.parse();
   }
 }
