@@ -12,11 +12,11 @@ import '../vector_drawable.dart';
 /// The AVD namespace.
 const String androidNS = 'http://schemas.android.com/apk/res/android';
 
-String getAttribute(
+String? getAttribute(
   List<XmlAttribute> attributes,
   String name, {
-  String def = '',
-  String namespace,
+  String? def = '',
+  String? namespace,
 }) {
   for (XmlAttribute attribute in attributes) {
     if (attribute.name.local == name) {
@@ -28,15 +28,15 @@ String getAttribute(
 
 /// Parses an AVD @android:viewportWidth and @android:viewportHeight attributes to a [Rect].
 DrawableViewport parseViewBox(List<XmlAttribute> el) {
-  final String rawWidth =
+  final String? rawWidth =
       getAttribute(el, 'viewportWidth', def: '', namespace: androidNS);
-  final String rawHeight =
+  final String? rawHeight =
       getAttribute(el, 'viewportHeight', def: '', namespace: androidNS);
   if (rawWidth == '' || rawHeight == '') {
     return const DrawableViewport(Size.zero, Size.zero);
   }
-  final double width = parseDouble(rawWidth);
-  final double height = parseDouble(rawHeight);
+  final double width = parseDouble(rawWidth)!;
+  final double height = parseDouble(rawHeight)!;
   return DrawableViewport(
     Size(width, height),
     Size(width, height),
@@ -44,20 +44,20 @@ DrawableViewport parseViewBox(List<XmlAttribute> el) {
 }
 
 Matrix4 parseTransform(List<XmlAttribute> el) {
-  final double rotation =
-      parseDouble(getAttribute(el, 'rotation', def: '0', namespace: androidNS));
+  final double rotation = parseDouble(
+      getAttribute(el, 'rotation', def: '0', namespace: androidNS))!;
   final double pivotX =
-      parseDouble(getAttribute(el, 'pivotX', def: '0', namespace: androidNS));
+      parseDouble(getAttribute(el, 'pivotX', def: '0', namespace: androidNS))!;
   final double pivotY =
-      parseDouble(getAttribute(el, 'pivotY', def: '0', namespace: androidNS));
-  final double scaleX =
+      parseDouble(getAttribute(el, 'pivotY', def: '0', namespace: androidNS))!;
+  final double? scaleX =
       parseDouble(getAttribute(el, 'scaleX', def: '1', namespace: androidNS));
-  final double scaleY =
+  final double? scaleY =
       parseDouble(getAttribute(el, 'scaleY', def: '1', namespace: androidNS));
   final double translateX = parseDouble(
-      getAttribute(el, 'translateX', def: '0', namespace: androidNS));
+      getAttribute(el, 'translateX', def: '0', namespace: androidNS))!;
   final double translateY = parseDouble(
-      getAttribute(el, 'translateY', def: '0', namespace: androidNS));
+      getAttribute(el, 'translateY', def: '0', namespace: androidNS))!;
 
   return Matrix4.identity()
     ..translate(pivotX, pivotY)
@@ -66,16 +66,16 @@ Matrix4 parseTransform(List<XmlAttribute> el) {
     ..translate(-pivotX + translateX, -pivotY + translateY);
 }
 
-DrawablePaint parseStroke(List<XmlAttribute> el, Rect bounds) {
-  final String rawStroke =
+DrawablePaint? parseStroke(List<XmlAttribute> el, Rect bounds) {
+  final String? rawStroke =
       getAttribute(el, 'strokeColor', def: null, namespace: androidNS);
   if (rawStroke == null) {
     return null;
   }
   return DrawablePaint(
     PaintingStyle.stroke,
-    color: parseColor(rawStroke).withOpacity(parseDouble(
-        getAttribute(el, 'strokeAlpha', def: '1', namespace: androidNS))),
+    color: parseColor(rawStroke)!.withOpacity(parseDouble(
+        getAttribute(el, 'strokeAlpha', def: '1', namespace: androidNS))!),
     strokeWidth: parseDouble(
         getAttribute(el, 'strokeWidth', def: '0', namespace: androidNS)),
     strokeCap: parseStrokeCap(el),
@@ -84,13 +84,13 @@ DrawablePaint parseStroke(List<XmlAttribute> el, Rect bounds) {
   );
 }
 
-double parseMiterLimit(List<XmlAttribute> el) {
+double? parseMiterLimit(List<XmlAttribute> el) {
   return parseDouble(
       getAttribute(el, 'strokeMiterLimit', def: '4', namespace: androidNS));
 }
 
 StrokeJoin parseStrokeJoin(List<XmlAttribute> el) {
-  final String rawStrokeJoin =
+  final String? rawStrokeJoin =
       getAttribute(el, 'strokeLineJoin', def: 'miter', namespace: androidNS);
   switch (rawStrokeJoin) {
     case 'miter':
@@ -105,7 +105,7 @@ StrokeJoin parseStrokeJoin(List<XmlAttribute> el) {
 }
 
 StrokeCap parseStrokeCap(List<XmlAttribute> el) {
-  final String rawStrokeCap =
+  final String? rawStrokeCap =
       getAttribute(el, 'strokeLineCap', def: 'butt', namespace: androidNS);
   switch (rawStrokeCap) {
     case 'butt':
@@ -119,21 +119,21 @@ StrokeCap parseStrokeCap(List<XmlAttribute> el) {
   }
 }
 
-DrawablePaint parseFill(List<XmlAttribute> el, Rect bounds) {
-  final String rawFill =
+DrawablePaint? parseFill(List<XmlAttribute> el, Rect bounds) {
+  final String? rawFill =
       getAttribute(el, 'fillColor', def: null, namespace: androidNS);
   if (rawFill == null) {
     return null;
   }
   return DrawablePaint(
     PaintingStyle.fill,
-    color: parseColor(rawFill)
-        .withOpacity(parseDouble(getAttribute(el, 'fillAlpha', def: '1'))),
+    color: parseColor(rawFill)!
+        .withOpacity(parseDouble(getAttribute(el, 'fillAlpha', def: '1'))!),
   );
 }
 
 PathFillType parsePathFillType(List<XmlAttribute> el) {
-  final String rawFillType =
+  final String? rawFillType =
       getAttribute(el, 'fillType', def: 'nonZero', namespace: androidNS);
   return rawFillType == 'nonZero' ? PathFillType.nonZero : PathFillType.evenOdd;
 }
