@@ -20,19 +20,19 @@ class DrawableAvdRoot extends DrawableRoot {
 
 /// An SVG Shape element that will be drawn to the canvas.
 class DrawableAvdPath extends DrawableShape {
-  const DrawableAvdPath(String id, Path path, DrawableStyle style)
+  const DrawableAvdPath(String? id, Path path, DrawableStyle style)
       : super(id, path, style);
 
   /// Creates a [DrawableAvdPath] from an XML <path> element
   factory DrawableAvdPath.fromXml(XmlElement el) {
     final String d =
-        getAttribute(el.attributes, 'pathData', def: '', namespace: androidNS);
+        getAttribute(el.attributes, 'pathData', def: '', namespace: androidNS)!;
     final Path path = parseSvgPathData(d);
-    assert(path != null);
+    assert(path != null); // ignore: unnecessary_null_comparison
 
     path.fillType = parsePathFillType(el.attributes);
-    final DrawablePaint stroke = parseStroke(el.attributes, path.getBounds());
-    final DrawablePaint fill = parseFill(el.attributes, path.getBounds());
+    final DrawablePaint? stroke = parseStroke(el.attributes, path.getBounds());
+    final DrawablePaint? fill = parseFill(el.attributes, path.getBounds());
 
     return DrawableAvdPath(
       getAttribute(el.attributes, 'id', def: ''),
@@ -62,16 +62,14 @@ Drawable parseAvdGroup(XmlElement el, Rect bounds) {
   for (XmlNode child in el.children) {
     if (child is XmlElement) {
       final Drawable el = parseAvdElement(child, bounds);
-      if (el != null) {
-        children.add(el);
-      }
+      children.add(el);
     }
   }
 
   final Matrix4 transform = parseTransform(el.attributes);
 
-  final DrawablePaint fill = parseFill(el.attributes, bounds);
-  final DrawablePaint stroke = parseStroke(el.attributes, bounds);
+  final DrawablePaint? fill = parseFill(el.attributes, bounds);
+  final DrawablePaint? stroke = parseStroke(el.attributes, bounds);
 
   return DrawableGroup(
     getAttribute(el.attributes, 'id', def: ''),
@@ -81,6 +79,6 @@ Drawable parseAvdGroup(XmlElement el, Rect bounds) {
       fill: fill,
       groupOpacity: 1.0,
     ),
-    transform: transform?.storage,
+    transform: transform.storage,
   );
 }
