@@ -11,13 +11,13 @@ import '../vector_drawable.dart';
 import 'colors.dart';
 import 'parsers.dart';
 
-double _parseRawWidthHeight(String raw) {
+double _parseRawWidthHeight(String? raw) {
   if (raw == '100%' || raw == '') {
     return double.infinity;
   }
   assert(() {
     final RegExp notDigits = RegExp(r'[^\d\.]');
-    if (!raw.endsWith('px') && raw.contains(notDigits)) {
+    if (!raw!.endsWith('px') && raw.contains(notDigits)) {
       print(
           'Warning: Flutter SVG only supports the following formats for `width` and `height` on the SVG root:\n'
           '  width="100%"\n'
@@ -27,7 +27,7 @@ double _parseRawWidthHeight(String raw) {
     }
     return true;
   }());
-  return double.tryParse(raw.replaceAll('px', '')) ?? double.infinity;
+  return double.tryParse(raw!.replaceAll('px', '')) ?? double.infinity;
 }
 
 /// Parses an SVG @viewBox attribute (e.g. 0 0 100 100) to a [Rect].
@@ -37,13 +37,13 @@ double _parseRawWidthHeight(String raw) {
 ///
 /// The [respectWidthHeight] parameter specifies whether `width` and `height` attributes
 /// on the root SVG element should be treated in accordance with the specification.
-DrawableViewport parseViewBox(
-  List<XmlEventAttribute> svg, {
+DrawableViewport? parseViewBox(
+  List<XmlEventAttribute>? svg, {
   bool nullOk = false,
 }) {
-  final String viewBox = getAttribute(svg, 'viewBox');
-  final String rawWidth = getAttribute(svg, 'width');
-  final String rawHeight = getAttribute(svg, 'height');
+  final String? viewBox = getAttribute(svg, 'viewBox');
+  final String? rawWidth = getAttribute(svg, 'width');
+  final String? rawHeight = getAttribute(svg, 'height');
 
   if (viewBox == '' && rawWidth == '' && rawHeight == '') {
     if (nullOk) {
@@ -67,7 +67,7 @@ DrawableViewport parseViewBox(
     );
   }
 
-  final List<String> parts = viewBox.split(RegExp(r'[ ,]+'));
+  final List<String> parts = viewBox!.split(RegExp(r'[ ,]+'));
   if (parts.length < 4) {
     throw StateError('viewBox element must be 4 elements long');
   }
@@ -75,26 +75,26 @@ DrawableViewport parseViewBox(
   return DrawableViewport(
     Size(width, height),
     Size(
-      parseDouble(parts[2]),
-      parseDouble(parts[3]),
+      parseDouble(parts[2])!,
+      parseDouble(parts[3])!,
     ),
     viewBoxOffset: Offset(
-      -parseDouble(parts[0]),
-      -parseDouble(parts[1]),
+      -parseDouble(parts[0])!,
+      -parseDouble(parts[1])!,
     ),
   );
 }
 
 /// Builds an IRI in the form of `'url(#id)'`.
-String buildUrlIri(List<XmlEventAttribute> attributes) =>
+String buildUrlIri(List<XmlEventAttribute>? attributes) =>
     'url(#${getAttribute(attributes, 'id')})';
 
 /// An empty IRI.
 const String emptyUrlIri = 'url(#)';
 
 /// Parses a `spreadMethod` attribute into a [TileMode].
-TileMode parseTileMode(List<XmlEventAttribute> attributes) {
-  final String spreadMethod =
+TileMode parseTileMode(List<XmlEventAttribute>? attributes) {
+  final String? spreadMethod =
       getAttribute(attributes, 'spreadMethod', def: 'pad');
   switch (spreadMethod) {
     case 'pad':
@@ -111,50 +111,51 @@ TileMode parseTileMode(List<XmlEventAttribute> attributes) {
 /// Parses an @stroke-dasharray attribute into a [CircularIntervalList].
 ///
 /// Does not currently support percentages.
-CircularIntervalList<double> parseDashArray(
-  List<XmlEventAttribute> attributes,
+CircularIntervalList<double>? parseDashArray(
+  List<XmlEventAttribute>? attributes,
 ) {
-  final String rawDashArray = getAttribute(attributes, 'stroke-dasharray');
+  final String? rawDashArray = getAttribute(attributes, 'stroke-dasharray');
   if (rawDashArray == '') {
     return null;
   } else if (rawDashArray == 'none') {
     return DrawableStyle.emptyDashArray;
   }
 
-  final List<String> parts = rawDashArray.split(RegExp(r'[ ,]+'));
+  final List<String> parts = rawDashArray!.split(RegExp(r'[ ,]+'));
   return CircularIntervalList<double>(
-      parts.map((String part) => parseDouble(part)).toList());
+      parts.map((String part) => parseDouble(part)!).toList());
 }
 
 /// Parses a @stroke-dashoffset into a [DashOffset].
-DashOffset parseDashOffset(List<XmlEventAttribute> attributes) {
-  final String rawDashOffset = getAttribute(attributes, 'stroke-dashoffset');
+DashOffset? parseDashOffset(List<XmlEventAttribute>? attributes) {
+  final String? rawDashOffset = getAttribute(attributes, 'stroke-dashoffset');
   if (rawDashOffset == '') {
     return null;
   }
 
-  if (rawDashOffset.endsWith('%')) {
+  if (rawDashOffset!.endsWith('%')) {
     final double percentage =
-        parseDouble(rawDashOffset.substring(0, rawDashOffset.length - 1)) / 100;
+        parseDouble(rawDashOffset.substring(0, rawDashOffset.length - 1))! /
+            100;
     return DashOffset.percentage(percentage);
   } else {
-    return DashOffset.absolute(parseDouble(rawDashOffset));
+    return DashOffset.absolute(parseDouble(rawDashOffset)!);
   }
 }
 
 /// Parses an @opacity value into a [double], clamped between 0..1.
-double parseOpacity(List<XmlEventAttribute> attributes) {
-  final String rawOpacity = getAttribute(attributes, 'opacity', def: null);
+double? parseOpacity(List<XmlEventAttribute>? attributes) {
+  final String? rawOpacity = getAttribute(attributes, 'opacity', def: null);
   if (rawOpacity != null) {
-    return parseDouble(rawOpacity).clamp(0.0, 1.0).toDouble();
+    return parseDouble(rawOpacity)!.clamp(0.0, 1.0).toDouble();
   }
   return null;
 }
 
 DrawablePaint _getDefinitionPaint(PaintingStyle paintingStyle, String iri,
     DrawableDefinitionServer definitions, Rect bounds,
-    {double opacity}) {
-  final Shader shader = definitions.getShader(iri, bounds);
+    {double? opacity}) {
+  final Shader? shader = definitions.getShader(iri, bounds);
   if (shader == null) {
     reportMissingDef(iri, '_getDefinitionPaint');
   }
@@ -167,22 +168,22 @@ DrawablePaint _getDefinitionPaint(PaintingStyle paintingStyle, String iri,
 }
 
 /// Parses a @stroke attribute into a [Paint].
-DrawablePaint parseStroke(
-  List<XmlEventAttribute> attributes,
-  Rect bounds,
+DrawablePaint? parseStroke(
+  List<XmlEventAttribute>? attributes,
+  Rect? bounds,
   DrawableDefinitionServer definitions,
-  DrawablePaint parentStroke,
+  DrawablePaint? parentStroke,
 ) {
-  final String rawStroke = getAttribute(attributes, 'stroke');
-  final String rawStrokeOpacity = getAttribute(
+  final String rawStroke = getAttribute(attributes, 'stroke')!;
+  final String? rawStrokeOpacity = getAttribute(
     attributes,
     'stroke-opacity',
     def: '1.0',
   );
-  final String rawOpacity = getAttribute(attributes, 'opacity');
-  double opacity = parseDouble(rawStrokeOpacity).clamp(0.0, 1.0).toDouble();
+  final String? rawOpacity = getAttribute(attributes, 'opacity');
+  double opacity = parseDouble(rawStrokeOpacity)!.clamp(0.0, 1.0).toDouble();
   if (rawOpacity != '') {
-    opacity *= parseDouble(rawOpacity).clamp(0.0, 1.0);
+    opacity *= parseDouble(rawOpacity)!.clamp(0.0, 1.0);
   }
 
   if (rawStroke.startsWith('url')) {
@@ -190,7 +191,7 @@ DrawablePaint parseStroke(
       PaintingStyle.stroke,
       rawStroke,
       definitions,
-      bounds,
+      bounds!,
       opacity: opacity,
     );
   }
@@ -201,16 +202,16 @@ DrawablePaint parseStroke(
     return DrawablePaint.empty;
   }
 
-  final String rawStrokeCap = getAttribute(attributes, 'stroke-linecap');
-  final String rawLineJoin = getAttribute(attributes, 'stroke-linejoin');
-  final String rawMiterLimit = getAttribute(attributes, 'stroke-miterlimit');
-  final String rawStrokeWidth = getAttribute(attributes, 'stroke-width');
+  final String? rawStrokeCap = getAttribute(attributes, 'stroke-linecap');
+  final String? rawLineJoin = getAttribute(attributes, 'stroke-linejoin');
+  final String? rawMiterLimit = getAttribute(attributes, 'stroke-miterlimit');
+  final String? rawStrokeWidth = getAttribute(attributes, 'stroke-width');
 
   final DrawablePaint paint = DrawablePaint(
     PaintingStyle.stroke,
     color: rawStroke == ''
         ? (parentStroke?.color ?? colorBlack).withOpacity(opacity)
-        : parseColor(rawStroke).withOpacity(opacity),
+        : parseColor(rawStroke)!.withOpacity(opacity),
     strokeCap: rawStrokeCap == 'null'
         ? parentStroke?.strokeCap ?? StrokeCap.butt
         : StrokeCap.values.firstWhere(
@@ -234,18 +235,18 @@ DrawablePaint parseStroke(
 }
 
 /// Parses a `fill` attribute.
-DrawablePaint parseFill(
-    List<XmlEventAttribute> el,
-    Rect bounds,
+DrawablePaint? parseFill(
+    List<XmlEventAttribute>? el,
+    Rect? bounds,
     DrawableDefinitionServer definitions,
-    DrawablePaint parentFill,
-    Color defaultFillColor) {
-  final String rawFill = getAttribute(el, 'fill');
-  final String rawFillOpacity = getAttribute(el, 'fill-opacity', def: '1.0');
-  final String rawOpacity = getAttribute(el, 'opacity');
-  double opacity = parseDouble(rawFillOpacity).clamp(0.0, 1.0).toDouble();
+    DrawablePaint? parentFill,
+    Color? defaultFillColor) {
+  final String rawFill = getAttribute(el, 'fill')!;
+  final String? rawFillOpacity = getAttribute(el, 'fill-opacity', def: '1.0');
+  final String? rawOpacity = getAttribute(el, 'opacity');
+  double opacity = parseDouble(rawFillOpacity)!.clamp(0.0, 1.0).toDouble();
   if (rawOpacity != '') {
-    opacity *= parseDouble(rawOpacity).clamp(0.0, 1.0);
+    opacity *= parseDouble(rawOpacity)!.clamp(0.0, 1.0);
   }
 
   if (rawFill.startsWith('url')) {
@@ -253,7 +254,7 @@ DrawablePaint parseFill(
       PaintingStyle.fill,
       rawFill,
       definitions,
-      bounds,
+      bounds!,
       opacity: opacity,
     );
   }
@@ -276,14 +277,14 @@ DrawablePaint parseFill(
   );
 }
 
-Color _determineFillColor(
-  Color parentFillColor,
+Color? _determineFillColor(
+  Color? parentFillColor,
   String rawFill,
   double opacity,
   bool explicitOpacity,
-  Color defaultFillColor,
+  Color? defaultFillColor,
 ) {
-  final Color color =
+  final Color? color =
       parseColor(rawFill) ?? parentFillColor ?? defaultFillColor;
   if (explicitOpacity && color != null) {
     return color.withOpacity(opacity);
@@ -293,32 +294,32 @@ Color _determineFillColor(
 }
 
 /// Parses a `fill-rule` attribute into a [PathFillType].
-PathFillType parseFillRule(List<XmlEventAttribute> attributes,
-    [String attr = 'fill-rule', String def = 'nonzero']) {
-  final String rawFillRule = getAttribute(attributes, attr, def: def);
+PathFillType? parseFillRule(List<XmlEventAttribute>? attributes,
+    [String attr = 'fill-rule', String? def = 'nonzero']) {
+  final String? rawFillRule = getAttribute(attributes, attr, def: def);
   return parseRawFillRule(rawFillRule);
 }
 
 /// Applies a transform to a path if the [attributes] contain a `transform`.
-Path applyTransformIfNeeded(Path path, List<XmlEventAttribute> attributes) {
-  final Matrix4 transform =
+Path? applyTransformIfNeeded(Path? path, List<XmlEventAttribute>? attributes) {
+  final Matrix4? transform =
       parseTransform(getAttribute(attributes, 'transform', def: null));
 
   if (transform != null) {
-    return path.transform(transform.storage);
+    return path!.transform(transform.storage);
   } else {
     return path;
   }
 }
 
 /// Parses a `clipPath` element into a list of [Path]s.
-List<Path> parseClipPath(
-  List<XmlEventAttribute> attributes,
+List<Path>? parseClipPath(
+  List<XmlEventAttribute>? attributes,
   DrawableDefinitionServer definitions,
 ) {
-  final String rawClipAttribute = getAttribute(attributes, 'clip-path');
+  final String? rawClipAttribute = getAttribute(attributes, 'clip-path');
   if (rawClipAttribute != '') {
-    return definitions.getClipPath(rawClipAttribute);
+    return definitions.getClipPath(rawClipAttribute!);
   }
 
   return null;
@@ -343,20 +344,20 @@ const Map<String, BlendMode> _blendModes = <String, BlendMode>{
 };
 
 /// Lookup the mask if the attribute is present.
-DrawableStyleable parseMask(
-  List<XmlEventAttribute> attributes,
+DrawableStyleable? parseMask(
+  List<XmlEventAttribute>? attributes,
   DrawableDefinitionServer definitions,
 ) {
-  final String rawMaskAttribute = getAttribute(attributes, 'mask');
+  final String? rawMaskAttribute = getAttribute(attributes, 'mask');
   if (rawMaskAttribute != '') {
-    return definitions.getDrawable(rawMaskAttribute);
+    return definitions.getDrawable(rawMaskAttribute!);
   }
 
   return null;
 }
 
 /// Parses a `font-weight` attribute value into a [FontWeight].
-FontWeight parseFontWeight(String fontWeight) {
+FontWeight? parseFontWeight(String? fontWeight) {
   if (fontWeight == null) {
     return null;
   }
@@ -390,11 +391,11 @@ FontWeight parseFontWeight(String fontWeight) {
 ///
 /// Remember that @style attribute takes precedence.
 DrawableStyle parseStyle(
-  List<XmlEventAttribute> attributes,
+  List<XmlEventAttribute>? attributes,
   DrawableDefinitionServer definitions,
-  Rect bounds,
-  DrawableStyle parentStyle, {
-  Color defaultFillColor,
+  Rect? bounds,
+  DrawableStyle? parentStyle, {
+  Color? defaultFillColor,
 }) {
   return DrawableStyle.mergeAndBlend(
     parentStyle,
@@ -429,6 +430,6 @@ DrawableStyle parseStyle(
         getAttribute(attributes, 'text-anchor', def: 'inherit'),
       ),
     ),
-    blendMode: _blendModes[getAttribute(attributes, 'mix-blend-mode')],
+    blendMode: _blendModes[getAttribute(attributes, 'mix-blend-mode')!],
   );
 }
