@@ -7,7 +7,7 @@ const String kXlinkNamespace = 'http://www.w3.org/1999/xlink';
 ///
 /// SVG 1.1 specifies that these attributes should be in the xlink namespace.
 /// SVG 2 deprecates that namespace.
-String getHrefAttribute(List<XmlEventAttribute> attributes) => getAttribute(
+String? getHrefAttribute(List<XmlEventAttribute>? attributes) => getAttribute(
       attributes,
       'href',
       namespace: kXlinkNamespace,
@@ -18,16 +18,16 @@ String getHrefAttribute(List<XmlEventAttribute> attributes) => getAttribute(
 /// is null or ''.
 ///
 /// Will look to the style first if it can.
-String getAttribute(
-  List<XmlEventAttribute> el,
+String? getAttribute(
+  List<XmlEventAttribute>? el,
   String name, {
-  String def = '',
-  String namespace,
+  String? def = '',
+  String? namespace,
   bool checkStyle = true,
 }) {
   String raw = '';
   if (checkStyle) {
-    final String style = _getAttribute(el, 'style')?.trim();
+    final String? style = _getAttribute(el!, 'style').trim();
     if (style != '' && style != null) {
       // Probably possible to slightly optimize this (e.g. use indexOf instead of split),
       // but handling potential whitespace will get complicated and this just works.
@@ -38,32 +38,30 @@ String getAttribute(
           orElse: () => '');
 
       if (raw != '') {
-        raw = raw.substring(raw.indexOf(':') + 1)?.trim();
+        raw = raw.substring(raw.indexOf(':') + 1).trim();
       }
     }
 
-    if (raw == '' || raw == null) {
-      raw = _getAttribute(el, name, namespace: namespace)?.trim();
+    if (raw == '') {
+      raw = _getAttribute(el, name, namespace: namespace).trim();
     }
   } else {
-    raw = _getAttribute(el, name, namespace: namespace)?.trim();
+    raw = _getAttribute(el!, name, namespace: namespace).trim();
   }
 
-  return raw == '' || raw == null ? def : raw;
+  return raw == '' ? def : raw;
 }
 
 String _getAttribute(
   List<XmlEventAttribute> list,
   String localName, {
   String def = '',
-  String namespace,
+  String? namespace,
 }) {
-  return list
-          .firstWhere(
-              (XmlEventAttribute attr) =>
-                  attr.name.replaceFirst('${attr.namespacePrefix}:', '') ==
-                  localName,
-              orElse: () => null)
-          ?.value ??
-      def;
+  for (XmlEventAttribute attr in list) {
+    if (attr.name.replaceFirst('${attr.namespacePrefix}:', '') == localName) {
+      return attr.value;
+    }
+  }
+  return def;
 }
