@@ -93,15 +93,15 @@ void main() {
     expect(parseTextAnchor('end'), DrawableTextAnchorPosition.end);
   });
 
-  T find<T extends Drawable>(Drawable drawable, String id) {
-    if (drawable?.id == id && drawable is T) {
+  T? find<T extends Drawable>(Drawable drawable, String id) {
+    if (drawable.id == id && drawable is T) {
       return drawable;
     }
 
     if (drawable is DrawableParent) {
       final DrawableParent parent = drawable;
-      for (Drawable item in parent.children) {
-        final Drawable found = find<T>(item, id);
+      for (Drawable item in parent.children!) {
+        final Drawable? found = find<T>(item, id);
 
         if (found != null) {
           return found as T;
@@ -126,10 +126,10 @@ void main() {
 
     expect(parseFontSize('larger'), parseFontSize('large'));
     expect(parseFontSize('larger', parentValue: parseFontSize('large')),
-        parseFontSize('large') * 1.2);
+        parseFontSize('large')! * 1.2);
     expect(parseFontSize('smaller'), parseFontSize('small'));
     expect(parseFontSize('smaller', parentValue: parseFontSize('large')),
-        parseFontSize('large') / 1.2);
+        parseFontSize('large')! / 1.2);
 
     expect(() => parseFontSize('invalid'),
         throwsA(const TypeMatcher<StateError>()));
@@ -199,10 +199,63 @@ void main() {
 
     final SvgParser parser = SvgParser();
     final DrawableRoot root = await parser.parse(svgStr);
-    expect(root.id.isEmpty, true);
+    expect(root.id!.isEmpty, true);
     expect(find<DrawableGroup>(root, 'Page-1') != null, true);
     expect(find<DrawableGroup>(root, 'iPhone-8') != null, true);
     expect(find<DrawableGroup>(root, 'stick_figure') != null, true);
     expect(find<DrawableShape>(root, 'Oval') != null, true);
+  });
+
+  test('Throws with unsupported elements with warnings as errors enabled',
+      () async {
+    const String svgStr = '''<?xml version="1.0" encoding="UTF-8"?>
+<svg width="27px" height="90px" viewBox="5 10 18 70" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+    <!-- Generator: Sketch 53 (72520) - https://sketchapp.com -->
+    <title>svg/stick_figure</title>
+    <desc>Created with Sketch.</desc>
+    <style> #Oval { fill: #D8D8D8; } </style>
+    <g id="Page-1" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
+        <g id="iPhone-8" transform="translate(-53.000000, -359.000000)" stroke="#979797">
+            <g id="stick_figure" transform="translate(53.000000, 359.000000)">
+                <ellipse id="Oval" cx="13.5" cy="12" rx="12" ry="11.5"></ellipse>
+                <path d="M13.5,24 L13.5,71.5" id="Line" stroke-linecap="square"></path>
+                <path d="M13.5,71.5 L1,89.5" id="Line-1" stroke-linecap="square"></path>
+                <path d="M13.5,37.5 L1,55.5" id="Line-2" stroke-linecap="square"></path>
+                <path d="M26.5,71.5 L14,89.5" id="Line-3" stroke-linecap="square" transform="translate(20.000000, 80.500000) scale(-1, 1) translate(-20.000000, -80.500000) "></path>
+                <path d="M26.5,37.5 L14,55.5" id="Line-2-Copy" stroke-linecap="square" transform="translate(20.000000, 46.500000) scale(-1, 1) translate(-20.000000, -46.500000) "></path>
+            </g>
+        </g>
+    </g>
+</svg>''';
+    final SvgParser parser = SvgParser();
+    expect(
+      parser.parse(svgStr, warningsAsErrors: true),
+      throwsA(anything),
+    );
+  });
+
+  test('Warns about unsupported elements by default', () async {
+    const String svgStr = '''<?xml version="1.0" encoding="UTF-8"?>
+<svg width="27px" height="90px" viewBox="5 10 18 70" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+    <!-- Generator: Sketch 53 (72520) - https://sketchapp.com -->
+    <title>svg/stick_figure</title>
+    <desc>Created with Sketch.</desc>
+    <style> #Oval { fill: #D8D8D8; } </style>
+    <g id="Page-1" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
+        <g id="iPhone-8" transform="translate(-53.000000, -359.000000)" stroke="#979797">
+            <g id="stick_figure" transform="translate(53.000000, 359.000000)">
+                <ellipse id="Oval" cx="13.5" cy="12" rx="12" ry="11.5"></ellipse>
+                <path d="M13.5,24 L13.5,71.5" id="Line" stroke-linecap="square"></path>
+                <path d="M13.5,71.5 L1,89.5" id="Line-1" stroke-linecap="square"></path>
+                <path d="M13.5,37.5 L1,55.5" id="Line-2" stroke-linecap="square"></path>
+                <path d="M26.5,71.5 L14,89.5" id="Line-3" stroke-linecap="square" transform="translate(20.000000, 80.500000) scale(-1, 1) translate(-20.000000, -80.500000) "></path>
+                <path d="M26.5,37.5 L14,55.5" id="Line-2-Copy" stroke-linecap="square" transform="translate(20.000000, 46.500000) scale(-1, 1) translate(-20.000000, -46.500000) "></path>
+            </g>
+        </g>
+    </g>
+</svg>''';
+
+    final SvgParser parser = SvgParser();
+    expect(await parser.parse(svgStr), isA<DrawableRoot>());
   });
 }
